Refactor library selection screens to use LibrarySelector component for improved code reuse and maintainability
